Fix problems with loading module options
With no top level 'sentry' object defined in nuxt-config.js, the code
would try to access property of undefined object as
this.options.sentry is undefined. Fixed by using deepmerge package which
both fixes the problem and makes code cleaner.

Also fixed problem with merging dsn property value with the object.
As dsn value is a string, don't use ... on it as that would turn string
into an array of letters.