Disable commit releasing by default since it's a breaking change
That change shouldn't have been released in minor version as it
is likely to break workflows since releasing commit often needs extra
configuration/work on the repo and/or sentry side.

For example for me it caused:
[error] Command failed: /var/lib/jenkins/jobs/workspace/node_modules/@sentry/cli/sentry-cli releases set-commits 2.22.107 --auto
error: API request failed
  caused by: sentry reported an error: You do not have permission to perform this action. (http status: 403)

And besides, this feature is badly documented, not even explaining what
`repo` is for.