Setup WebpackPlugin during webpack:config hook
The `webpack:config` can be used to manipulate the list of webpack build that will be performed. If a custom config is added by another module, that other config would not have the `WebpackPlugin` added to its plugins since `extendBuild` is called only when NuxtJS build the initial list of `webpackConfigs`.