fix: infer "release" automatically if "config.release" not specified
Use Sentry CLI to automatically deduce "release" configuration if not
specified manually.

It's necessary to store the release in a generated file at build-time,
and import that file at run-time, so that the server-side Sentry can use
it on initialization (which happens within the module itself rather than
within a server-side plugin).

Resolves #200