fix: initialize server-side Sentry when using Nuxt programmatically
Switched from "listen" to "ready" hook so that server-side Sentry is
also initialized when starting Nuxt programmatically from within a server
framework like expressjs.

The side effect of using "ready" hook is that Sentry will also be
initialized on building which seems like an OK thing to do.

Also, in a typical case, "ready" triggers before "listen" which should
actually be better as it would also catch issues with starting server
listener.

Resolves #230