fix!: use different hook for initializing server-side Sentry instance (#403)
BREAKING CHANGE: The server-side `process.sentry` will be created slightly later than before WHEN running the "build" action. It will be created before the pages are built (on `build:compile` hook) while before it was available a bit earlier on `ready` hook (with an issue that it was not always able to pass the project version to Sentry).